requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch
torchvision
albumentations
ensemble-boxes
numpy
matplotlib
pandas
Pillow
opencv-python